Rethink function introduction, close #5
The function introduction is now baked into the introduction to R
programming, and the lecture on R packages is move to the same day
as the programming basics. This freed up a whole day that now is
moved to after the tidyverse/tidy data day for a general catch-up
before moving on to other topics.